First pass at adding content
This content is just a brain dump -- it needs to be formatted such that ideas about the different tools are formatted in a table or at least a separate list, defined by the categories I've determined here. It would also be great to add links to the web sites of the various tools.